Refactor into seperate functions
